+++ packages/web-components/keen-slider-lazy-images/src/components/index.ts
@@ -1,0 +1,1 @@
+export * as KeenSliderLazyImages from "./keen-slider-lazy-images.js";